requirements.txt
python-telegram-bot==20.3
aiosqlite==0.19.0